Only kill the JVM if it ran all the tasks
Summary:
Since the JVMs get stuck sometimes after running tasks we
decided to start killing them after the task is finished. However it
gets in the way of the JVM reuse logic.
This is failing the tests.

Test Plan:
ant test (actually ant test succeeds on my box anyway, but it
fails in the automated system)

Reviewers: aching, rvadali

Reviewed By: rvadali

+++ src/mapred/org/apache/hadoop/mapred/JvmManager.java
@@ -216,7 +216,11 @@ class JvmManager {
           jvmRunner.taskRan();
           // This JVM is done. Unfortunately sometimes the
           // process hangs around, so we should sigkill it
-          jvmRunner.kill();
+          // However it is only applicable to the JVMs that
+          // ran all the tasks they were supposed to
+          if (jvmRunner.ranAll()) {
+            jvmRunner.kill();
+          }
         }
       }
     }
